requirements.txt
paho-mqtt==2.1.0
influxdb-client==1.45.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
slowapi==0.1.9
python-dotenv==1.0.1
